backend/requirements.txt
Django==5.0
mysqlclient==2.2.0
djangorestframework==3.14.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-django==0.45b0
opentelemetry-instrumentation-mysqlclient==0.45b0
